Update project from VS2013 to VS2015

+++ src/Object.cpp
@@ -734,7 +734,7 @@ void Object::serialize(Table& table)
     robj.set("position",getPosition());
     robj.set("rotation",getRotation().valueDegrees());
     robj.set("scale",getScale());
-    robj.set("z", getZ());
+    robj.set("z", (float)getZ());
     robj.set("canDrawChilds", (float)getCanDrawChilds());
     
     if(parent)
@@ -762,8 +762,8 @@ void Object::deserialize(const Table& table)
     setPosition(table.getVector2D("position",getPosition()));
     setRotation(Angle( Degree( table.getFloat("rotation",   getRotation().valueDegrees() ) ) ) );
     setScale(table.getVector2D("scale",      getScale()));
-    setZ(table.getFloat("z", getZ()));
-    setCanDrawChilds(table.getFloat("canDrawChilds", getCanDrawChilds()) != 0.0f);
+    setZ(table.getFloat("z", (float)getZ()));
+    setCanDrawChilds(table.getFloat("canDrawChilds", (float)getCanDrawChilds()) != 0.0f);
 
     if(table.existsAsType("parentMode",Table::STRING))
     {
